requirements.txt
torch
torchaudio
transformers
datasets[audio]
soundfile
librosa
fastapi
uvicorn
sounddevice
soundfile
numpy
requests
webrtcvad
streamlit
numpy
webrtcvad
requests
python-dotenv
openai>=1.40.0
wavio
streamlit-webrtc